[GH-1] change argument parsing to use getopt_long()
The cmd parser is still compatible with the previous argument structure,
so there is not change in the usage itself.

As a result of this change, main no longer needs to check the return
value of handle_args(), and thus the return value has been removed and
error handling has been unified within handle_args() itself.

Include headers are now sorted alphabetically.